requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-appconf==1.0.1
django-compressor==1.5
django-ipware==1.1.0
django-storages==1.1.8
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.4.8.1
gunicorn==19.3.0
oauthlib==0.7.2
Pillow==2.8.2
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
